requirements.txt
Flask==2.3.2
openai
polygon-api-client